+++ Mediapipe.Net/Framework/Packet/SidePacket.cs
@@ -32,7 +32,7 @@ namespace Mediapipe.Net.Framework.Packet
             GC.KeepAlive(this);
 
             // Oh gosh²... the Activator²...
-            return (T?)Activator.CreateInstance(typeof(T), packetPtr, true);
+            return (T?)Activator.CreateInstance(typeof(T), (IntPtr)packetPtr, true);
         }
 
         public void Emplace<T>(string key, Packet<T> packet)
